src/front/requirements.txt
uvicorn~=0.12.3
dominate~=2.6.0
fastapi~=0.62.0
aiofiles~=0.6.0
azure-identity 
azure-keyvault 
azure-mgmt-resource
starlette~=0.13.6
msrestazure~=0.6.4
msal~=1.8.0